# src/WoodgroveGroceriesApi/OpenAPI.yaml
openapi: 3.0.3
info:
  title: Woodgrove Groceries API
  description: API for managing products, carts, and checkout/payment operations in the Woodgrove Groceries demo.
  version: 1.0.0
servers:
  - url: https://api.woodgrovegroceries.com/v1
    description: Main API server
...
security:
  - oauth2: []
paths:
...
  /checkout:
    post:
      tags:
        - Checkout
      summary: Perform checkout for a cart
      operationId: checkoutCart
      description: Initiates the checkout process for a given cart, calculating the total and preparing the order.
      security:
        - oauth2: ['Checkout.Process']
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutRequest'
      responses:
        '200':
          description: Checkout details, including total and order status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutResponse'
        '400':
          description: Bad request (e.g., empty cart)
        '401':
          description: Unauthorized
        '404':
          description: Cart not found
  /checkout/{orderId}/pay:
    post:
      tags:
        - Checkout
      summary: Pay for an order
      operationId: payOrder
      description: Processes the payment of the order generated during checkout.
      parameters:
        - name: orderId
          in: path
          required: true
          description: ID of the order to pay
          schema:
            type: string
      security:
        - oauth2: ['Checkout.Process']
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
      responses:
        '200':
          description: Payment successfully processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          description: Bad request (invalid payment data)
        '401':
          description: Unauthorized
        '402':
          description: Payment error (insufficient funds, etc.)
        '404':
          description: Order not found
components:
  securitySchemes:
    oauth2:
      type: oauth2
      description: OAuth 2.0 authentication using Microsoft Entra ID
      flows:
        implicit:
          authorizationUrl: https://lagomarciamdemo2.ciamlogin.com/c92a6719-2559-47df-baa3-c9f02517c42c/oauth2/v2.0/authorize
          scopes:
            'Products.Read': Read access to product information
            'Products.Write': Write access to product information
            'Carts.Read': Read access to cart information
            'Carts.Write': Write access to cart information
            'Checkout.Process': Access to checkout and payment processing
...
  schemas:
...
    CheckoutRequest:
      type: object
      properties:
        cartId:
          type: string
          description: ID of the cart to be processed
        address:
          type: string
          description: Shipping address
      required:
        - cartId
    CheckoutResponse:
      type: object
      properties:
        orderId:
          type: string
          description: ID of the generated order
        cartId:
          type: string
          description: ID of the associated cart
        totalAmount:
          type: number
          format: float
          description: Total amount to be paid
        currency:
          type: string
          description: Currency used
        status:
          type: string
          description: Order status (e.g., "PENDING_PAYMENT")
      required:
        - orderId
        - cartId
        - totalAmount
        - currency
        - status
    PaymentRequest:
      type: object
      properties:
        paymentMethod:
          type: string
          description: Payment method (credit card, PayPal, etc.)
        cardNumber:
          type: string
          description: Card number (if applicable)
        expirationDate:
          type: string
          description: Card expiration date (mm/yy) (if applicable)
        cvv:
          type: string
          description: Card security code (if applicable)
      required:
        - paymentMethod
    PaymentResponse:
      type: object
      properties:
        orderId:
          type: string
          description: ID of the paid order
        paymentStatus:
          type: string
          description: Result of the payment (e.g., "PAID", "FAILED")
        transactionId:
          type: string
          description: Transaction ID from the payment processor (if applicable)
      required:
        - orderId
        - paymentStatus